backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
bcrypt
python-dotenv
google-generativeai
httpx
